refactor(combinatorics): isolate difference-set contracts
